Move NPC utils to new class
