+++ src/main/java/io/github/cccm5/util/NPCUtil.java
@@ -1,24 +1,12 @@
-package io.github.cccm5;
+package io.github.cccm5.util;
 
 import net.citizensnpcs.api.npc.NPC;
 import net.citizensnpcs.api.npc.NPCRegistry;
 import net.citizensnpcs.api.trait.Trait;
-import net.countercraft.movecraft.craft.Craft;
-import net.countercraft.movecraft.MovecraftLocation;
-import net.countercraft.movecraft.util.hitboxes.HitBox;
-import org.bukkit.Location;
-import org.bukkit.Material;
-import org.bukkit.World;
-import org.bukkit.inventory.Inventory;
-import org.bukkit.inventory.InventoryHolder;
-import org.bukkit.inventory.ItemStack;
 
 import java.util.ArrayList;
-import java.util.List;
 
-
-public class Utils
-{
+public class NPCUtil {
     public static ArrayList<NPC> getNPCsWithTrait(Class<? extends Trait> c){
         ArrayList<NPC> npcs = new ArrayList<>();
         for(NPCRegistry registry : net.citizensnpcs.api.CitizensAPI.getNPCRegistries())
@@ -27,7 +15,4 @@ public class Utils
                     npcs.add(npc);
         return npcs;
     }
-    
-    
-
 }
